TB-68 Fix 블로그, 숙소 정보에 아이디 할당

+++ src/main/java/com/move/TripBalance/result/Blog.java
@@ -14,13 +14,13 @@ public class Blog {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private int id;
     private String title;
     private String contents;
     private String url;
     private String blogName;
     private String thumbnail;
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
